Create log streams lazily (#97)
Assume log streams exist when emitting the log, this avoids making a
request to create the log stream when a logger emits an event after the
process starts up.

In my use case, Python processes are spawned to react to an event, e.g.
a file was uploaded to S3. On the first upload, the log stream must be
created. For all subsequent uploads however, the log stream can be
assumed to exist.

Also, when spawning multiple processes quickly, e.g. in reaction to many
files being uploaded, AWS throttles the create_log_stream operation,
replying with:

```
ClientError in <my_process>
An error occurred (ThrottlingException) when calling the CreateLogStream
operation (reached max retries: 4): Rate exceeded
```

That response is not handled by the logger and crashes the Python process.